tech-enhancements/tools/smart-notes-app/requirements.txt
Flask==3.0.0
google-generativeai==0.4.0
python-dotenv==1.0.0